Modularized functions and variables out of main.js
Created audio_module.js, constants.js, animation_polyfill.js, and player_js folder containing related modules
