+++ Music_Visualization_Application/static/js/main.js
@@ -1,8 +1,3 @@
-////////////////////////////////////////Import Statements////////////////////////////////////////
-
-// Import Howler Library for audio playback handling
-import "./howler_js/howler.js";
-
 // Import the SoundCloud client_id associated with your application
 import { client_id } from "./config.js";
 
@@ -10,121 +5,57 @@ import { client_id } from "./config.js";
 import {
   displayErrorMessage,
   clearErrorMessage,
-} from "./error_message_module.js";
+} from "./player_js/error_message_module.js";
 
 // Import the requestTrack function
-import { requestTrack } from "./soundcloud_request_module.js";
+import { requestTrack } from "./player_js/soundcloud_request_module.js";
 
 // Import the function that updates what track is being played, along with the information displayed about that track
 import {
   displayTimeElapsed,
   displayProgressBarElapsed,
   updateMusicPlayer,
-} from "./update_track_module.js";
+} from "./player_js/update_track_module.js";
 
 // Import the functions to change the icon associated with the middle playback control button to reflect Pause, Play, and Loading.
 import {
   displayPauseButton,
   displayPlayButton,
   displayLoadingButton,
-} from "./update_buttons_module.js";
+} from "./player_js/update_buttons_module.js";
 
 import { World } from "./three_js/World.js";
 
+import { animation_polyfill } from "./animation_polyfill.js";
+
+import {
+  canvasContainer,
+  INPUT_FORM,
+  PLAY_BUTTON,
+  VOLUME_UP_BUTTON,
+  VOLUME_DOWN_BUTTON,
+  DEFAULT_TRACK,
+  FREQUENCY_SAMPLESIZE,
+} from "./constants.js";
+
+import {
+  globalAudio,
+  globalAnalyser,
+  globalDataArray,
+  getSecondsElapsed,
+  getPercentageElapsed,
+} from "./audio_module.js";
 ////////////////////////////////////////////////////////////////////////////////////////
 ("use strict");
 
-// requestAnimationFrame polyfill by Erik Möller.
-// Fixes from Paul Irish, Tino Zijdel, Andrew Mao, Klemen Slavic, Darius Bacon and Joan Alba Maldonado.
-// Adapted from https://gist.github.com/paulirish/1579671 which derived from
-// http://paulirish.com/2011/requestanimationframe-for-smart-animating/
-// http://my.opera.com/emoller/blog/2011/12/20/requestanimationframe-for-smart-er-animating
-// Added high resolution timing. This window.performance.now() polyfill can be used: https://gist.github.com/jalbam/cc805ac3cfe14004ecdf323159ecf40e
-// MIT license
-// Gist: https://gist.github.com/jalbam/5fe05443270fa6d8136238ec72accbc0
-(function () {
-  var vendors = ["webkit", "moz", "ms", "o"],
-    vp = null;
-  for (
-    var x = 0;
-    x < vendors.length &&
-    !window.requestAnimationFrame &&
-    !window.cancelAnimationFrame;
-    x++
-  ) {
-    vp = vendors[x];
-    window.requestAnimationFrame =
-      window.requestAnimationFrame || window[vp + "RequestAnimationFrame"];
-    window.cancelAnimationFrame =
-      window.cancelAnimationFrame ||
-      window[vp + "CancelAnimationFrame"] ||
-      window[vp + "CancelRequestAnimationFrame"];
-  }
-  if (
-    /iP(ad|hone|od).*OS 6/.test(window.navigator.userAgent) ||
-    !window.requestAnimationFrame ||
-    !window.cancelAnimationFrame
-  ) {
-    //iOS6 is buggy.
-    var lastTime = 0;
-    window.requestAnimationFrame = function (callback, element) {
-      var now = window.performance.now();
-      var nextTime = Math.max(lastTime + 16, now);
-      return setTimeout(function () {
-        callback((lastTime = nextTime));
-      }, nextTime - now);
-    };
-    window.cancelAnimationFrame = clearTimeout;
-  }
-})();
-
-////////////////////////////////////////Constants////////////////////////////////////////
-
-// Capture the form HTML element associated with the submission form for the SoundCloud URL by the user
-const canvasContainer = document.getElementById("animation_div");
-const INPUT_FORM = document.getElementById("user_form_element");
-// Capture the button HTML element associated with the play/pause button and volume up and volume down buttons.
-const PLAY_BUTTON = document.getElementById("play_pause_button");
-const VOLUME_UP_BUTTON = document.getElementById("volume_up_button");
-const VOLUME_DOWN_BUTTON = document.getElementById("volume_down_button");
-// Initialize the default track
-const DEFAULT_TRACK =
-  "https://ia802807.us.archive.org/23/items/cd_mozart_wolfgang-amadeus-mozart-richard-goode/disc1/07.%20Wolfgang%20Amadeus%20Mozart%20-%20Rondo%20in%20A%20minor%2C%20K.%20511_sample.mp3";
-// The quantity of fast fourier transform samples of the track's frequency to be sampled per call to the Analyzer Node.
-const FREQUENCY_SAMPLESIZE = 256;
-
-////////////////////////////////////////Global variables/objects////////////////////////////////////////
-
-// Initialize new Howler.js object with its default track and behaviors
-let globalAudio = new Howl({
-  src: [DEFAULT_TRACK],
-  volume: 0.5,
-  autoplay: false,
-  html5: false,
-  loop: false,
-});
-
-// Modify Howler.js object's inner AudioContext to include an Analyser node which will allow us to retrieve data from the song being played.
-let globalAnalyser = Howler.ctx.createAnalyser();
-Howler.masterGain.connect(globalAnalyser);
-
-// Each time globalAnalyser.getByteFrequencyData() is called and passed the array globalDataArray as an argument, the array will contain frequency data of the song at the time of the function call.
-globalAnalyser.fftSize = FREQUENCY_SAMPLESIZE;
-let bufferLength = globalAnalyser.frequencyBinCount;
-let globalDataArray = new Uint8Array(bufferLength);
+// Polyfill requestAnimationFrame for smoother animation
+animation_polyfill();
 
 // A default array with 0s in all its indices that is fed into the very first render of the animation when the application is first loaded and we have no array from actual song data yet.
 let cleanDataArray = [];
 for (let i = 0; i < FREQUENCY_SAMPLESIZE / 2; i++) {
   cleanDataArray[i] = 0;
 }
-
-// Boolean flag indicating whether or not we want to be collecting frequency data from the song.
-let collectingTrackFrequencies = false;
-
-////////////////////////////////////////Global function definitions////////////////////////////////////////
-
-// Only display the Play button on the middle music player control button when the track has been fully loaded.
 
 // Load a new instanced threeJS world and render it using the cleanDataArray. After the animation has loaded once, we will use actual array data containing frequency information held in globalDataArray. We do this because render current requires an array argument.
 let world = new World(canvasContainer);
@@ -141,7 +72,7 @@ function callPerFrame() {
   displayProgressBarElapsed(percentageElapsed);
 
   // If we intend to collect frequency data during this frame, do so through our analyser node and store the resulting array of data in the globalDataArray
-  if (collectingTrackFrequencies) {
+  if (globalAudio.collectingTrackFrequencies) {
     globalAnalyser.getByteFrequencyData(globalDataArray);
 
     // Render a single three.js frame that is informed by data
@@ -152,55 +83,6 @@ function callPerFrame() {
 }
 
 callPerFrame();
-
-// When audio has loaded, for safety ensure that frequency collection flag is set to false since no music is playing until the user presses play. As such we display the play button.
-globalAudio.on("load", () => {
-  pauseFrequencyCollection();
-  displayPlayButton();
-});
-
-// When audio is playing, set the frequency collection flag to true and show the pause button to indicate the song can be paused.
-globalAudio.on("play", () => {
-  resumeFrequencyCollection();
-  displayPauseButton();
-});
-
-// When a song finishes playing "on end", change the icon on the middle button of the music player to display a Play icon. If play is clicked after this occurs, song playback will begin again from the beginning of the track.
-globalAudio.on("end", () => {
-  pauseFrequencyCollection();
-  displayPlayButton();
-});
-
-// This function sets the flag for whether or not we want to collect frequency data for what track is currently playing to FALSE.
-function pauseFrequencyCollection() {
-  collectingTrackFrequencies = false;
-}
-
-// This function sets the flag for whether or not we want to collect frequency data for what track is currently playing to TRUE.
-function resumeFrequencyCollection() {
-  collectingTrackFrequencies = true;
-}
-
-// Based on the current track's playback progress, return the number of seconds that have elapsed during the track's playback.
-// This snippet was taken from the Howler.js documentation's music player example found here: https://github.com/goldfire/howler.js/blob/master/examples/player/player.js
-function getSecondsElapsed() {
-  let secondsElapsed = globalAudio.seek() || 0;
-  return secondsElapsed;
-}
-
-// Given the number of seconds elapsed during the track's playback, convert it to a percentage value relative to the total duration (in seconds) of the track.
-// This snippet was taken from the Howler.js documentation's music player example found here: https://github.com/goldfire/howler.js/blob/master/examples/player/player.js
-function getPercentageElapsed(secondsElapsed) {
-  let percentageElapsed = (secondsElapsed / globalAudio.duration()) * 100 || 0;
-  return percentageElapsed;
-}
-
-// Make a GET request to the Flask server, sending the track id of the requested song in the request. The Flask server will create its own get request and obtain the URL of the direct mp3 stream/file of the track itself and return it.
-// Once the URL has been obtained, we change the source of the globalAudio howler object to be the URL of the direct mp3 stream for the requested song. We update the middle button of the music player to display a loading icon while the
-// request is being made, and to also display a play icon when the globalAudio source has been updated and is ready to play.
-function loadTrack(SoundCloud_track) {
-  globalAudio.changeSrc(SoundCloud_track.streamSource);
-}
 
 ////////////////////////////////////////HTML Event Listeners////////////////////////////////////////
 
@@ -223,13 +105,13 @@ INPUT_FORM.addEventListener("submit", function (event) {
   const results = requestTrack(user_url, client_id)
     .then((SoundCloud_track) => {
       // Pause collection of frequency data
-      pauseFrequencyCollection();
+      globalAudio.pauseFrequencyCollection();
 
       // Stop playing audio
       globalAudio.stop();
 
       // Load the new track
-      loadTrack(SoundCloud_track);
+      globalAudio.changeSrc(SoundCloud_track.streamSource);
 
       // Update the Music Player with the new track's information, such as artist, title, total duration.
       updateMusicPlayer(SoundCloud_track);
@@ -252,11 +134,11 @@ INPUT_FORM.addEventListener("submit", function (event) {
 PLAY_BUTTON.addEventListener("click", function (event) {
   if (globalAudio.playing()) {
     globalAudio.pause();
-    pauseFrequencyCollection();
+    globalAudio.pauseFrequencyCollection();
     displayPlayButton();
   } else if (globalAudio.state() === "loaded") {
     globalAudio.play();
-    resumeFrequencyCollection();
+    globalAudio.resumeFrequencyCollection();
     displayPauseButton();
   }
 });
