+++ Music_Visualization_Application/static/js/player_js/update_buttons_module.js
